kernel: ext4 module handles ext{2,3} filesystems.
From now on, the ext4 kernel module handles ext2 and ext3 formatted
filesystem as well. ext2 and ext3 filesystems remain in their format
and are NOT converted to ext4.
But we will be able to drop the ext2 and ext3 modules which will
decrease the footprint size of the kernel, because ext4 code handles
the legacy filesystems, too.