kernel: Update to Linux 3.2.
Introduce new way of creating configurations:

From now on, there are no configuration files as they come
out of "make oldconfig" for every single architecture and kernel
version.

There is now a system that creates those configurations
and so it is easier to track changes, see differences between
architectures and a lot more.

There are configuration files for armv5tel kernels but they
do not build because of problems with the grsecurity patchset.