filesystem: Add /var/{lock,run} symlinks and fix scriplet.
systemd apparently does not bind-mount /var/lock and /var/run
anymore but symlinks them instead to their corresponding directories
in /run.

The scriptlet contained an error which caused a crash on all
architectures except x86_64 because of a bad checking of
the directory existance. Hopefully this is now properly fixed.