icap: Changed config file directory.
There is a dependency when building the packages, the modules seem
to rely on the master sysconfdir and ignoring there own sysconfdir.
This is whierd when building the modules with the old icap master
packages leading to config files in /etc even when you point it
elsewhere. You need to build the icap packages first and point the
sysconfdir to /etc/anywhere and then the modules will install there
too, ignoring what you´ve configured with the sysconfdir for the
package.

Thats why both will install in /etc/c-icap because you can not link
the modules elsewhere.

Fixes #10125