ulogd2: Update to version 2.0.5
This package adds logging of the flows and packets to an sqlite
database that can be used to get statistical output.

Logging to syslog as known from IPFire 2 has been added as well
but is only temporary until there is a way to view the logged
packets on the console.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>