lighttpd: Drop package
lighttpd was originally used to run the IPFire web infrastructure
and also inteded to be used to host the web user interface in
IPFire. It was chosen because of its lightweight design and
small footprint.

However, development of lighttpd has stalled since a long time
and it does not support all the major web technologies any more.

Hence it is now removed from the distribution in favor of other
alternatives.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>