openvpn: Drop package
The IPFire Developer Summit decided to drop OpenVPN support
from IPFire 3 since it has many issues that don't seem to get
solved in the near future and nobody volunteered to maintain
this package.

Development is focussing on IPsec for now and OpenVPN may
come back later when resources have been freed or there
is enough demand (including volunteers to work on it).

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>