bison: Update to 3.0.4.
This is a major update to the latest stable version of bison.

A short excerpt of the changelog:

Several deprecated features were removed, such as support for YYFAIL, yystype,
yyltype, YYLEX_PARAM, and YYPARSE_PARAM. The user epilogue is no longer affected
by internal #defines. The recently-introduced dependency on stdio.h when
locations are enabled was removed. Caret errors, as introduced in 2.7, are now
activated by default. The "-Werror" option was enhanced, and categorization of
warnings and errors was modified. Many other changes and various bugfixes were
made.

I also added flex as a build dependency to prevent from errors when running
the testsuite.

Fixes #10403.

Signed-off-by: Stefan Schantl <stefan.schantl@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>