traceroute: Update to 2.1
The main significant change is new adaptive algorithm for waiting.
Traditional traceroute implementation always waited the whole timeout (default 5.0 second) for any probe.
But if we already have some replies from the same hop, or even from some next hop, we can use the round trip time of such a reply as a hint to determine the actual reasonable amount of time to wait.

Signed-off-by: Alexander Marx <alexander.marx@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>