Update nmap to 7.30
This patch updates nmap to version 7.30 and introduce a new package
nmap-ncat.

The nmap project provides since version 7.0 a netcat implementation
which is ipv4 and ipv6 able.

Netcat is a dependency of libvirt and to avoid unnecessary dependencies
the tool ncat is shipped in an extra package.
The symlink nc point to the ncat implementation and is necessary because
libvirt use /usr/bin/nc to execute netcat.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>