glibc: Compile with -O2
Instead of compiling with -O3, we now compile this with -O2
which is the default for the whole system. It looks like binaries
launch and perform a little bit slower than on some other
distributions, but there is no explanation to it.

This change will reduce the size of the libc and sibling libraries
which will hopefully allow to load them faster. Code might potentially
execute slower on some machines, but this impact might be less than
the startup time of the binaries.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>