kernel: Fix compiled-in version number
When we have a .0 release of the kernel, the internal kernel
versioning is 4.20.0, but the tarball is called 4.20 only.

This causes the kernel to look for its modules in the wrong place.

This patch overwrites the compiled-in version number so that
it is in sync with the directory of the installed modules.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>