clamav: Disable llvm support
When enabled, LLVM provides the capability to just-in-time compile ClamAV bytecode signatures.
Without LLVM, ClamAV uses a built-in bytecode interpreter to execute bytecode signatures.
The mechanism is different, but the results are same and the performance overall is comparable.

At present only LLVM versions up to LLVM 3.6.2 are supported by ClamAV, and LLVM 3.6.2 is old enough
that newer distributions no longer provide it. Therefore, the ClamAV Devs recommend using the
--disable-llvm configure option.

Fixes #11601

Signed-off-by: Stefan Schantl <stefan.schantl@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>